% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_balance.R
\name{check_balance}
\alias{check_balance}
\title{Testing Covariate Balance}
\usage{
check_balance(data, treatment, num = NULL, cat = NULL, round = NULL)
}
\arguments{
\item{data}{Data frame containing the variables}

\item{treatment}{Name of treatment/exposure variable (must be binary)}

\item{num}{Vector of numeric variable names (e.g., c("age", "income"))}

\item{cat}{Vector of categorical variable names (e.g., c("sex", "race"))}

\item{round}{Number of decimal places for rounding numeric outputs
(default: NULL for no rounding)}
}
\value{
A list
}
\description{
Testing Covariate Balance
}
\examples{
check_balance(data = lalonde,
              treatment = "treat",
              num = c("age", "educ"),
              cat = c("race", "married"),
              round = 3)
}
