% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_balance.R
\name{plot_balance}
\alias{plot_balance}
\title{Plot Covariate Balance}
\usage{
plot_balance(
  data,
  treatment,
  num = NULL,
  cat = NULL,
  numeric_plot = "histogram"
)
}
\arguments{
\item{data}{Data frame used in check_balance.}

\item{treatment}{Treatment variable used in check_balance (must be binary).}

\item{num}{Vector of numeric variable names (e.g., c("age", "educ"))}

\item{cat}{Vector of categorical variable names (e.g., c("race", "married"))}

\item{numeric_plot}{Either "histogram" or "boxplot" for numeric covariates
(default "histogram).}
}
\value{
A list of ggplot objects: $numeric, $categorical
}
\description{
Creates side-by-side plots for numeric and categorical covariates after
running check_balance.
Numeric covariates can be displayed as histograms or boxplots.
Categorical covariates are displayed as bar plots
}
\examples{
# After running check_balance
plot_balance(data = lalonde,
             treatment = "treat",
             num = c("age", "educ"),
             cat = c("race", "married"),
             numeric_plot = "histogram")

}
